AST: Split off ProtocolConformanceRef.cpp from ProtocolConformance.cpp
