+++ lib/AST/CMakeLists.txt
@@ -74,6 +74,7 @@ add_swift_host_library(swiftAST STATIC
   PlatformKind.cpp
   PrettyStackTrace.cpp
   ProtocolConformance.cpp
+  ProtocolConformanceRef.cpp
   RawComment.cpp
   RequirementEnvironment.cpp
   RequirementMachine/ConcreteContraction.cpp
